refactor: split RoundMap into map hooks + layers
