+++ apps/web/src/components/round/map/markerFactories.ts
@@ -1,0 +1,174 @@
+import { mapboxgl } from '../../../lib/mapbox'
+
+export const MARKER_COLORS = {
+  tee: '#1C211C',
+  approach: '#A66A1F',
+  green: '#1F3D2C',
+  putt: '#1F3D2C',
+  ball: '#1F3D2C',
+  pin: '#A33A2A',
+} as const
+
+// ---------------------------------------------------------------------------
+// DOM marker factories — keep them lightweight, no React per marker.
+//
+// Mapbox writes inline `transform: translate3d(...)` on the marker's
+// outer element to position it, so we can't put our own transform there
+// (it would be wiped on the next pan). Each factory returns an outer
+// (positioned by Mapbox) plus an inner `content` we own — that's where
+// hover scale and drag glow are applied. See `attachDragFx`.
+// ---------------------------------------------------------------------------
+
+export interface MarkerParts {
+  outer: HTMLElement
+  content: HTMLElement
+}
+
+export function makeNumberedMarker(
+  n: number,
+  fill: string,
+  text: string,
+): MarkerParts {
+  const outer = document.createElement('div')
+  outer.style.display = 'flex'
+  outer.style.alignItems = 'center'
+  outer.style.justifyContent = 'center'
+  const content = document.createElement('div')
+  content.style.width = '24px'
+  content.style.height = '24px'
+  content.style.borderRadius = '999px'
+  content.style.background = fill
+  content.style.color = text
+  content.style.fontFamily = 'Inter, sans-serif'
+  content.style.fontWeight = '600'
+  content.style.fontSize = '12px'
+  content.style.display = 'flex'
+  content.style.alignItems = 'center'
+  content.style.justifyContent = 'center'
+  content.style.border = '2px solid #FBF8F1'
+  content.style.transition =
+    'transform 120ms ease, box-shadow 120ms ease'
+  content.textContent = String(n)
+  outer.appendChild(content)
+  return { outer, content }
+}
+
+export function makeAimMarker(): HTMLElement {
+  const outer = document.createElement('div')
+  outer.style.display = 'flex'
+  outer.style.alignItems = 'center'
+  outer.style.justifyContent = 'center'
+  const dot = document.createElement('div')
+  dot.style.cssText = [
+    'width:14px',
+    'height:14px',
+    'border-radius:999px',
+    'background:#A66A1F',
+    'border:2px solid #FBF8F1',
+    'pointer-events:none',
+  ].join(';')
+  outer.appendChild(dot)
+  outer.title = 'Aim point'
+  return outer
+}
+
+export function makeDistancePill(label: string): HTMLElement {
+  const el = document.createElement('div')
+  el.style.cssText = [
+    'background:rgba(28,33,28,0.85)',
+    'color:#F2EEE5',
+    'font-family:JetBrains Mono, monospace',
+    'font-size:11px',
+    'font-weight:500',
+    'letter-spacing:0.04em',
+    'padding:3px 8px',
+    'border-radius:999px',
+    'pointer-events:none',
+    'white-space:nowrap',
+  ].join(';')
+  el.textContent = label
+  return el
+}
+
+export function makeIconMarker(
+  label: string,
+  bg: string,
+  fg: string,
+): MarkerParts {
+  const outer = document.createElement('div')
+  const content = document.createElement('div')
+  content.style.padding = '3px 6px'
+  content.style.background = bg
+  content.style.color = fg
+  content.style.fontFamily = 'JetBrains Mono, monospace'
+  content.style.fontSize = '9px'
+  content.style.letterSpacing = '0.14em'
+  content.style.fontWeight = '500'
+  content.style.border = `1px solid ${fg}`
+  content.style.borderRadius = '2px'
+  content.style.transition =
+    'transform 120ms ease, box-shadow 120ms ease'
+  content.textContent = label
+  outer.appendChild(content)
+  return { outer, content }
+}
+
+export interface FlagParts extends MarkerParts {
+  flag: HTMLElement
+}
+
+export function makeFlagMarker(color: string): FlagParts {
+  const outer = document.createElement('div')
+  const content = document.createElement('div')
+  content.style.width = '16px'
+  content.style.height = '24px'
+  content.style.position = 'relative'
+  content.style.transition =
+    'transform 120ms ease, box-shadow 120ms ease'
+  const pole = document.createElement('div')
+  pole.style.cssText =
+    'position:absolute;left:6px;top:0;width:2px;height:24px;background:#FBF8F1'
+  const flag = document.createElement('div')
+  flag.style.cssText = `position:absolute;left:8px;top:1px;width:9px;height:7px;background:${color};transition:background 120ms ease`
+  const base = document.createElement('div')
+  base.style.cssText =
+    'position:absolute;left:5px;top:22px;width:4px;height:2px;border-radius:1px;background:#FBF8F1'
+  content.appendChild(pole)
+  content.appendChild(flag)
+  content.appendChild(base)
+  outer.appendChild(content)
+  return { outer, content, flag }
+}
+
+export function attachDragFx(opts: {
+  outer: HTMLElement
+  content: HTMLElement
+  marker: mapboxgl.Marker
+  tooltip: string
+  onDragColor?: (active: boolean) => void
+}) {
+  const { outer, content, marker, tooltip, onDragColor } = opts
+  outer.title = tooltip
+  outer.style.cursor = 'grab'
+  let dragging = false
+  outer.addEventListener('mouseenter', () => {
+    if (!dragging) content.style.transform = 'scale(1.2)'
+  })
+  outer.addEventListener('mouseleave', () => {
+    if (!dragging) content.style.transform = ''
+  })
+  marker.on('dragstart', () => {
+    dragging = true
+    outer.style.cursor = 'grabbing'
+    content.style.transform = 'scale(1.2)'
+    content.style.boxShadow = '0 0 0 4px rgba(166,106,31,0.55)'
+    onDragColor?.(true)
+  })
+  marker.on('dragend', () => {
+    dragging = false
+    outer.style.cursor = 'grab'
+    content.style.transform = ''
+    content.style.boxShadow = ''
+    onDragColor?.(false)
+  })
+}
